refactor(cli): split main.py into commands/ subpackage (phase 1)
Mechanical move only: main.py 2,181 -> 80 lines, nine command modules
under commands/, with main.py keeping the Typer app, the version-sync
literals, and __main__. Commands register explicitly rather than
decorating the app global, which would otherwise create an import cycle.

Patch targets are repointed in the same commit and pinned by
tests/test_cli_patch_targets.py. Verified mechanical independently: AST
comparison shows all 49 top-level defs byte-identical with none added,
missing, or body-changed, and the 34-command help surface is
JSON-identical to the pre-split module.
