+++ tests/test_cli_drain.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from unittest.mock import patch
 
 import netllm_cli.main as cli_main
+from netllm_cli.commands import observe as cli_observe
 from netllm_core.models import NetllmConfig, save_config
 from typer.testing import CliRunner
 
@@ -50,7 +51,7 @@ class _FakeClient:
 
 def test_drain_on_posts_draining_true(tmp_path: Path) -> None:
     cfg_path = _cfg_path(tmp_path)
-    with patch.object(cli_main.httpx, "Client", _FakeClient):
+    with patch.object(cli_observe.httpx, "Client", _FakeClient):
         result = runner.invoke(cli_main.app, ["drain", "on", "--config", str(cfg_path)])
     assert result.exit_code == 0, result.output
     assert "Draining" in result.output
@@ -61,7 +62,7 @@ def test_drain_on_posts_draining_true(tmp_path: Path) -> None:
 
 def test_drain_off_posts_draining_false(tmp_path: Path) -> None:
     cfg_path = _cfg_path(tmp_path)
-    with patch.object(cli_main.httpx, "Client", _FakeClient):
+    with patch.object(cli_observe.httpx, "Client", _FakeClient):
         result = runner.invoke(
             cli_main.app, ["drain", "off", "--config", str(cfg_path)]
         )
@@ -73,7 +74,7 @@ def test_drain_off_posts_draining_false(tmp_path: Path) -> None:
 
 def test_drain_defaults_to_on(tmp_path: Path) -> None:
     cfg_path = _cfg_path(tmp_path)
-    with patch.object(cli_main.httpx, "Client", _FakeClient):
+    with patch.object(cli_observe.httpx, "Client", _FakeClient):
         result = runner.invoke(cli_main.app, ["drain", "--config", str(cfg_path)])
     assert result.exit_code == 0, result.output
     _url, body = _FakeClient.last_post
@@ -104,6 +105,6 @@ class _UnreachableClient:
 
 def test_drain_agent_unreachable_fails_cleanly(tmp_path: Path) -> None:
     cfg_path = _cfg_path(tmp_path)
-    with patch.object(cli_main.httpx, "Client", _UnreachableClient):
+    with patch.object(cli_observe.httpx, "Client", _UnreachableClient):
         result = runner.invoke(cli_main.app, ["drain", "on", "--config", str(cfg_path)])
     assert result.exit_code != 0
